# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@kinde.com
    name: Kinde Support Team
    url: https://kinde.com/docs/
  description: Provides endpoints to manage your Kinde Businesses
  termsOfService: https://kinde.com/docs/important-information/terms-of-service/
  title: Kinde Management API
  version: "1"
servers:
- url: "https://{businessName}.kinde.com/"
  variables:
    businessName:
      default: app
      description: Business name created in the Kinde admin area.
security:
- kindeBearerAuth: []
...
paths:
  /oauth2/user_profile:
    get:
      description: |
        Contains the id, names and email of the currently logged in user.
      operationId: getUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user_profile'
          description: Details of logged in user V1.
        "403":
          description: Invalid credentials.
      summary: Get User Profile
      tags:
      - OAuth
      x-accepts: application/json
      x-tags:
      - tag: OAuth
  /oauth2/v2/user_profile:
    get:
      description: |
        Contains the id, names, profile picture URL and email of the currently logged in user.
      operationId: getUserProfileV2
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user_profile_v2'
          description: Details of logged in user V2.
        "403":
          description: Invalid credentials.
      summary: Returns the details of the currently logged in user
      tags:
      - OAuth
      x-accepts: application/json
      x-tags:
      - tag: OAuth
...
components:
  schemas:
...
    user_profile:
      example:
        preferred_email: preferred_email
        last_name: last_name
        id: id
        provided_id: provided_id
        first_name: first_name
        picture: picture
      properties:
        id:
          description: Unique id of the user in Kinde.
          type: string
        preferred_email:
          description: Default email address of the user in Kinde.
          type: string
        provided_id:
          description: Value of the user's id in a third-party system when the user
            is imported into Kinde.
          nullable: true
          type: string
        last_name:
          description: User's last name.
          type: string
        first_name:
          description: User's first name.
          type: string
        picture:
          description: URL that point's to the user's picture or avatar
          type: string
      type: object
    user_profile_v2:
      example:
        sub: sub
        updated_at: 0
        name: name
        id: id
        given_name: given_name
        provided_id: provided_id
        family_name: family_name
        email: email
        picture: picture
      properties:
        id:
          description: Unique id of the user in Kinde (deprecated).
          type: string
        sub:
          description: Unique id of the user in Kinde.
          type: string
        provided_id:
          description: Value of the user's id in a third-party system when the user
            is imported into Kinde.
          nullable: true
          type: string
        name:
          description: Users's first and last name separated by a space.
          type: string
        given_name:
          description: User's first name.
          type: string
        family_name:
          description: User's last name.
          type: string
        updated_at:
          description: Date the user was last updated at (In Unix time).
          type: integer
        email:
          description: User's email address if available.
          type: string
        picture:
          description: URL that point's to the user's picture or avatar
          type: string
      type: object
...
  securitySchemes:
    kindeBearerAuth:
      bearerFormat: JWT
      description: |
        Accessing the API requires obtaining an access token when login in through Kinde. However, management functions (for e.g. user management) requires an access token
        obtained using the client_credentials flow.
      scheme: bearer
      type: http